requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
todoist-python==8.1.2
typing==3.7.4.3
urllib3==1.25.10
